comment.bigint

+++ db/schema.rb
@@ -81,7 +81,7 @@ ActiveRecord::Schema[8.0].define(version: 2026_04_26_094103) do
   end
 
   create_table "sessions", force: :cascade do |t|
-    t.integer "user_id", null: false
+    t.bigint "user_id", null: false
     t.string "ip_address"
     t.string "user_agent"
     t.datetime "created_at", null: false
@@ -90,7 +90,7 @@ ActiveRecord::Schema[8.0].define(version: 2026_04_26_094103) do
   end
 
   create_table "spots", force: :cascade do |t|
-    t.integer "user_id", null: false
+    t.bigint "user_id", null: false
     t.string "name", null: false
     t.integer "category", null: false
     t.string "address"
